// temp_frontend/src/lib/models/AuthenticationResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Données de la réponse
 */
export type AuthenticationResponse = {
    id?: string;
    email?: string;
    role?: string;
    firstName?: string;
    lastName?: string;
    photoUrl?: string;
    city?: string;
    university?: string;
    specialization?: string;
    grade?: string;
    subjects?: Array<string>;
    certification?: string;
    registrationDate?: string;
    lastLogin?: string;
    token?: string;
};