// temp_frontend/src/lib/models/StudentRegisterRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StudentRegisterRequest = {
    email: string;
    password?: string;
    confirmPassword?: string;
    firstName?: string;
    lastName?: string;
    photoUrl?: string;
    city?: string;
    university?: string;
    specialization?: string;
};