// src/lib/models/TeacherRegisterRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TeacherRegisterRequest = {
    email: string;
    password?: string;
    confirmPassword?: string;
    firstName?: string;
    lastName?: string;
    photoUrl?: string;
    city?: string;
    university?: string;
    grade?: string;
    subjects?: Array<string>;
    certification?: string;
};